--- TUho/apps/Noticias/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-from .models import Noticias
-
-# Register your models here.
-
-admin.site.register(Noticias)
-
-
