--- TUho/apps/Noticias/apps.py
@@ -1,8 +1,0 @@
-from django.apps import AppConfig
-
-
-class NoticiasConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'apps.Noticias'
-
-
