--- TUho/apps/Noticias/urls.py
@@ -1,26 +1,0 @@
-"""
-URL configuration for config project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/5.0/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-from django.urls import path
-from apps.Noticias import views
-from django.contrib.auth.decorators import login_required
-
-urlpatterns = [
-    path('Noticias/',login_required (views.NoticiasView), name="Noticias"),
-    path('CrearNoticia/',login_required (views.CrearNoticia), name="CrearNoticia"),
-    path('EditarNoticia/<int:id>/',login_required (views.EditarNoticia), name="EditarNoticia"),
-    path('EliminarNoticia/<int:id>/',login_required (views.EliminarNoticia), name="EliminarNoticia"),
-]
