+++ TUho/apps/Usuarios/urls.py
@@ -20,5 +20,14 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth import views as auth_views
 
 urlpatterns = [
+    path('Login/', views.Login, name="Login"),
+    path('Registrar/', views.Registrar, name="Registrar"),
+    path('CerrarSesion/',login_required(views.CerrarSesion), name="CerrarSesion"), 
+    # Contraseña Olvidada
+    path('reset_password/', views.RestablecerContraseña.as_view() , name="password_reset"),
+    path('reset_password_send/', views.RestablecerContraseñaConfirmado, name="password_reset_done"),
+    path('reset/<uidb64>/<token>', views.CambiarContraseña.as_view(), name="password_reset_confirm"),
+    path('reset_password_complete/', views.CambiarContraseñaConfirmado, name="password_reset_complete"),
+     # Informacion Personal
     path('ActualizarInf/', views.ActualizarInf, name="ActualizarInf"),
 ]
