+++ TUho/config/settings.py
@@ -38,9 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'apps.Plataforma',
-    'apps.Autenticacion',
     'apps.Usuarios',
-    'apps.Noticias',
 ]
 
 MIDDLEWARE = [
