+++ TUho/static/css/datatable.css
@@ -7,23 +7,32 @@
 }
 
 .card-title{
-    color: #005296;
+    color: #0d6efd;
 }
 
 .card-header{
-    background: #fff !important;
+    background:#fff!important;
+    width: 100%;
 }
 
 input{
     box-shadow: none !important;
 }
 
-.iconos{
+.iconos {
     color:black;
+    font-size: 1rem;
+}
+
+.iconos i{
+    color:black;
+    font-size: 1.1rem;
 }
 
 .iconos:hover{
-    font-size: 1.3rem;
+    color: #0d6efd;
+}
+.iconos i:hover{
     color: #0d6efd;
 }
