+++ TUho/static/js/Modals Inicio.js
@@ -1,0 +1,9 @@
+var Inicio = document.getElementById('Inicio')
+Inicio.addEventListener('show.bs.modal', function (event) {
+  // Button that triggered the modal
+  var button = event.relatedTarget
+  // Extract info from data-bs-* attributes
+  var recipient = button.getAttribute('data-bs-whatever')
+  
+})
+
