+++ TUho/static/js/Modals.js
@@ -1,11 +1,3 @@
-var Inicio = document.getElementById('Inicio')
-Inicio.addEventListener('show.bs.modal', function (event) {
-  // Button that triggered the modal
-  var button = event.relatedTarget
-  // Extract info from data-bs-* attributes
-  var recipient = button.getAttribute('data-bs-whatever')
-  
-})
 
 var exampleModal = document.getElementById('exampleModal')
 if (exampleModal != null) {
