+++ TUho/static/js/Trim registrar.js
@@ -2,6 +2,7 @@ var username = document.getElementById("username");
 var password1 = document.getElementById("password1");
 var password2 = document.getElementById("password2");
 var email = document.getElementById("email");
+var boton =  document.getElementById("boton");
 
 
   const validate_space_trim = () => {
@@ -15,3 +16,4 @@ username.addEventListener("input",validate_space_trim);
 password1.addEventListener("input",validate_space_trim);
 password2.addEventListener("input",validate_space_trim);
 email.addEventListener("input",validate_space_trim);
+boton.addEventListener("click", c, false)
