+++ TUho/static/js/Trim.js
@@ -1,17 +1,39 @@
-var ci = document.getElementById("inputCi");
-var nombre = document.getElementById("inputName");
-var tel = document.getElementById("inputNt");
-var email = document.getElementById("inputEmail");
+var username = document.getElementById("username");
+var password = document.getElementById("password");
+var boton = document.getElementById("boton");
 
 
-  const validate_space_trim = () => {
-    ci.value = ci.value.trim();
-    nombre.value = nombre.value.trim();
-    tel.value = tel.value.trim();
-    email.value = email.value.trim();
+const errorContainer = document.querySelector("#error-container")
+        const createMessage = (message) => {
+            return `
+            <div class="alert alert-danger alert-dismissible fade show" role="alert"
+            style="height:40px; position: fixed; right: 20px; top: 50px; display: flex; align-items: center; padding-right: 0rem;">
+            ${message}
+            <button style="font-size: 10px; border-bottom: none; position: relative; box-shadow: none;" type="button"
+            class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+            </button>
+            </div>
+            `
+        }
+
+const validate_space_trim = () => {
+  username.value = username.value.trim();
+  password.value = password.value.trim();
+}
+
+const c = (e)=>{
+  if(username.value == ""){
+    errorContainer.innerHTML = createMessage("El nombre de ususario no es válido")
+    e.preventDefault()
   }
+  if(password.value == ""){
+    errorContainer.innerHTML = createMessage("La contraseña no es válida valido")
+    e.preventDefault()
+ 
 
-ci.addEventListener("input",validate_space_trim);
-nombre.addEventListener("input",validate_space_trim);
-tel.addEventListener("input",validate_space_trim);
-email.addEventListener("input",validate_space_trim);
+}
+}
+
+username.addEventListener("input",validate_space_trim);
+password.addEventListener("input",validate_space_trim);
+boton.addEventListener("click", c, false)
