+++ TUho/templates/Base.html
@@ -185,7 +185,7 @@
     {% endblock %}
     <script src=" {% static 'libs/Bootstrap/js/bootstrap.bundle.min.js' %}"></script>
     <script src=" {% static 'js/script.js' %}"></script>
-    <script src="{% static 'js/Modals.js' %}"></script>
+    <script src="{% static 'js/Modals Inicio.js' %}"></script>
     <!-- End Container Js -->
