+++ TUho/templates/Plataforma/Cambiar Rol.html
@@ -1,0 +1,66 @@
+{% extends "Base Admin.html" %}
+{% load static %}
+
+<!-- Conatiner CSS -->
+{% block css %}
+<link rel="stylesheet" href=" {% static 'css/Crear Noticia.css' %}">
+{% endblock %}
+<!-- End Conatiner CSS -->
+
+{% block content %}
+<div class="container-form">
+    <div class="form-rol">
+        <div class="contact-info">
+            <h3 class="title">Cambiar rol </h3>
+            <p class="text">
+                Como administrador de la plataforma, tienes la capacidad de asignar diferentes roles a 
+                los usuarios para adaptarse a sus necesidades y responsabilidades.
+                Recuerda que cada rol tiene sus propias características y permisos, por lo que 
+                es importante seleccionar el que mejor se ajuste a tus requerimientos.
+            </p>
+        </div>
+        <div class="contact-form">
+            <span class="circle one"></span>
+            <span class="circle two"></span>
+            <form action="."  method="POST">
+                {% csrf_token %}
+                <h3 class="title">Cambiar</h3>  
+                {% if response %}
+                {% if response == 'correcto' %}
+                <div class="alert alert-success alert-dismissible fade show" role="alert"
+                style="height:40px; position: fixed; right: 20px; top: 50px; display: flex; align-items: center; padding-right: 0rem;">
+                {{ message }}
+                <button style="font-size: 10px; border-bottom: none; position: relative; box-shadow: none;" type="button"
+                    class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+                </button>
+                </div>
+                {% endif %}
+                {% endif %}
+                <div class="input-container select-rol">
+                    <input type="hidden" name="role">
+                    <div class="input selector">
+                    </div>
+                    <label for="">Rol de Usuario</label>
+                    <span>Rol de Usuario</span>
+                  </div>
+                  <div class="select-options" name="role">
+                    <div aria-valuetext="Administración" value="Administración" class="element">Administración</div>
+                    <div aria-valuetext="Supervisor" value="Supervisor" class="element">Supervisor</div>
+                    <div aria-valuetext="Usuario" value="Usaurio" class="element">Usuario</div>
+                  </div>
+                  <div class="btn">
+                    <a href="{% url 'Usuarios' %}" class="btn2">Cancelar</a>
+                    <input type="submit" value="Aceptar" class="btn2">
+                  </div>
+            </form>
+        </div>
+    </div>
+</div>
+{% endblock %}
+
+<!-- Conatiner JS -->
+{% block js %}
+<script src="{% static 'js/FormularioSelectRol.js' %}"></script>
+
+{% endblock %}
+<!-- End Conatiner JS -->
