+++ TUho/templates/Plataforma/Informacion Personal.html
@@ -48,8 +48,8 @@
                 <div class="input-container focus">
                     <input type="text" name="name" value="{{ request.user.carnet }}" class="input" id="inputCi"
                         disabled>
-                    <label for="">Carnet ID</label>
-                    <span>Carnet ID</span>
+                    <label for="">Carnet</label>
+                    <span>Carnet</span>
                 </div>
                 <div class="input-container focus">
                     <input type="email" name="email" value="{{ request.user.email }}" class="input" id="inputEmail"
