+++ TUho/templates/Plataforma/Mis Trámites.html
@@ -14,12 +14,13 @@
 {% block content %}
 <main class="content-inside px-3 py-2 mx-2">
     <section>
-        <div class="card p-2">
+        <div class="card">
             <div class="card-header">
                 <h5 class="card-title">
                     Mis Trámites
                 </h5>
             </div>
+            <div class="p-2">
             <table id="Tramites" class="display table table-striped table-hover table-primary" style="width:100%">
                 <thead class="encabezado">
                     <tr>
@@ -40,6 +41,7 @@
                     </tr>
                 </tbody>
             </table>
+            </div>
         </div>
     </section>
 </main>
