+++ TUho/templates/Plataforma/Noticias.html
@@ -37,7 +37,7 @@
     {% else %}
     <section class="d-flex flex-column justify-content-center align-content-center">
         <p class="text text-center">
-            No hay noticias nuevas hechas recientemente
+            No hay noticias hechas recientemente
         </p>
         <div class="d-flex justify-content-center mb-2 mt-1 ">
             <a href="{% url 'CrearNoticia' %}" class="btn btn-primary">Crear Noticia</a>
