+++ TUho/templates/Plataforma/Tramites.html
@@ -8,7 +8,6 @@
 <link rel="stylesheet" href=" {% static 'libs/Datatable/CSS/buttons.bootstrap5.css' %}">
 <link rel="stylesheet" href=" {% static 'libs/Datatable/CSS/responsive.bootstrap5.css' %}">
 <link rel="stylesheet" href=" {% static 'css/datatable.css' %}">
-<link rel="stylesheet" href=" {% static 'libs/Chart/css/chart.css' %}">
 {% endblock %}
 <!-- End Conatiner CSS -->
 
@@ -16,12 +15,13 @@
 {% block content %}
 <main class="content-inside px-3 py-2">
     <section id="tablaTramites">
-        <div class="card p-2">
+        <div class="card">
             <div class="card-header">
                 <h5 class="card-title">
                     Trámites
                 </h5>
             </div>
+            <div class="p-2">
             <table id="Tramites" class="display table table-striped table-hover table-primary"
                 style="width:100%; height: 100%;">
                 <thead class="encabezado">
@@ -49,10 +49,11 @@
                             </select></td>
                         <td style="text-align: center; font-size: 1.2rem;"><a class="iconos p-2" href="#"><i
                                     class="bx bx-trash"></i></a><a class="iconos p-2" href="#"><i
-                                    class="bi bi-eye"></i></a></td>
+                                    class="bx bx-info-circle"></i></a></td>
                     </tr>
                 </tbody>
             </table>
+            </div>
         </div>
     </section>
 </main>
